requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blockbuster==1.5.25
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
forbiddenfruit==0.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-core==0.3.75
langchain-openai==0.3.32
langgraph==0.6.6
langgraph-api==0.4.7
langgraph-checkpoint==2.1.1
langgraph-cli==0.4.0
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.9.0
langgraph-sdk==0.2.4
langsmith==0.4.21
openai==1.102.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.47.3
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
truststore==0.10.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
xxhash==3.5.0
zstandard==0.24.0
